[Step2] feature(lotto): 로또 화면 구성에 대한 책임 분리 (LottoInputView, LottoOutputView)
